fix issue with initializing admin sdk, also some comment fixes
this was a hard issue to track down because it only occurs when
two functions that require/initialize the admin sdk occur on the
same "warm" cloud functions instance.

to resolve the issue, a custom module initializing/exporting all the admin
functionalities was written. Any funcaionalities, are then imported by
the functions using them. This prevents admin.initializeApp() from running
more than once.